سلعتي/backend/requirements.txt
coverage==6.0
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
# Faker==8.12.1
# faker-commerce==1.0.3
Pyrebase==3.0.27
python-dateutil==2.8.2
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
sqlparse==0.4.1
